A utility function to identify integer columns in a data frame

The function is used to suggest integer data types imported as double from CSV files. It is used to create an internal list of 'known integer columns' that are used to force integer data types when reading CSV files from GenX/PowerGenome files.

Usage

mip_suggest_integer_columns(
  data,
  min_value = -100,
  max_value = 100,
  threshold = 1,
  print_level = 1
)

Arguments

data

data frame, data table, or arrow query result

min_value

minimum value for a column to be considered an integer

max_value

maximum value for a column to be considered an integer

threshold

minimum proportion of integer values in a column (default 1.0 - all values)

print_level

integer indicating the level of verbosity (0 = silent, 1 = print integer columns)

Value

character vector of column names identified as integers